Move folder structure to a self-contained component
