+++ components/query-builder/utils.ts
@@ -1,4 +1,6 @@
 import { Step, InputType, Column } from "./types";
+import { clsx, type ClassValue } from "clsx";
+import { twMerge } from "tailwind-merge";
 
 const defaultComparators = ["equals", "contains", "begins with", "ends with"];
 const stringComparators = ["equals", "contains", "begins with", "ends with"];
@@ -75,5 +77,9 @@ const getPropsPerStep = (
       };
   }
 };
+
+export function cn(...inputs: ClassValue[]) {
+  return twMerge(clsx(inputs));
+}
 
 export { getPropsPerStep };
